Renamed protobuf_utils to protobuf_decoder and removed write functionality
