+++ lutris/util/amazon/sds_proto2.py
@@ -1,4 +1,4 @@
-from lutris.util.amazon.protobuf_utils import (
+from lutris.util.amazon.protobuf_decoder import (
     Message, type_bool, type_bytes, type_enum, type_int64, type_string, type_uint32
 )
